Upgrade packages, add UNSAFE_componentWillMount prefix, refactor CRA scripts
