+++ package.json
@@ -43,9 +43,9 @@
     "prop-types": "^15.7.2",
     "pubsub-js": "^1.7.0",
     "rc-tree": "2.1.0",
-    "react": "16.8.6",
+    "react": "^16.8.6",
     "react-color": "^2.18.0",
-    "react-dom": "16.8.6",
+    "react-dom": "^16.8.6",
     "react-scripts": "^2.1.8",
     "short-number": "^1.0.6",
     "store": "^2.0.12",
